src/requirements.txt
numpy
matplotlib
spatialmath-python
roboticstoolbox-python
